Potential fix for #34 and #45
Added a condition in Loop for DrivesThatCanHoldMoreThanJob;
After a job is started on a DriveThatCanHoldMoreThanJob, we query Get-PlotoTempDrives to see if we have another Drive available. If so, we call Invoke-PlotoJob again with all params. Please note: This is NOT tested AT ALL! Can't test it on work machine, as I they won't let me download and install chia.exe :P