Refactor core logic of Invoke-PlotoJob
I just realized i was maintaining copies of code that added complexity for nothing. 
The core logic was basically revamped and the codebased heavily simplified. It now does not wait for the amount specified in -WaitTimeBetweenPlotOnSameDisks and lock itself in there. It periodically  checks if there a disk for plotting available and picks the one with the least jobs in progress. You can define the intervall in Start-PlotoSpawns. I use 5. If there is a disk, it checks if this disk has any jobs in progress that were spawned within the lock period of the duration defined in -WaitTimeBetweenPlotOnSameDisks. if not it spawns a job, if yes, it checks the next until all drives have been iterated trough. It then checks again in the defined intervall you specified.